requirements.txt
Django==1.6.4
Pillow==2.4.0
django-appconf==0.6
django-crispy-forms==1.4.0
django-frontend==1.1.0
django-frontend-skeleton==3.0.0
django-geoposition==0.2.0
django-gmapi==1.0.1
django-leaflet==0.13.2
six==1.6.1
